requirements.txt
streamlit
qrcode
gTTS
PyMuPDF
Pillow
pytesseract